Create CAT1 Theory G2 Slot Q3
A digital locker in the bank is protected with a security mechanism. To open the
locker a password of 9 characters is required. The input characters should be
accepted as 3x3 matrix and two diagonal characters of the matrix are concatenated
(refer to the example given below) and compared with the password already stored
in a character array for authentication. Write a C program to implement this logic
for password verification.